+++ app/Http/Controllers/CategoryController.php
@@ -129,7 +129,10 @@ class CategoryController extends Controller
     public function delete(Category $category,$id)
     {
         $data = Category::find($id);
+        if($data->image && Storage::disk('public')->exists($data->image))
+        {
         Storage::delete($data->image);
+        }
         $data->delete();
         return redirect('admin/category');
     }
